Joined legacy ChannelMessagesScreen with new Inbox screen
