+++ lib/constants/app_constants.dart
@@ -7,6 +7,7 @@ part 'enums.dart';
 part 'identifiers.dart';
 part 'numerical.dart';
 part 'paths.dart';
+part 'routes.dart';
 part 'schemes.dart';
 part 'styles.dart';
 part 'themes.dart';
