+++ lib/widgets/screens/inbox/inbox_invites_received.dart
@@ -3,7 +3,8 @@ import 'dart:math';
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:mm_flutter_app/constants/app_constants.dart';
-import 'package:mm_flutter_app/widgets/atoms/inbox_list_tile.dart';
+import 'package:mm_flutter_app/utilities/debug_logger.dart';
+import 'package:mm_flutter_app/widgets/molecules/inbox_list_tile.dart';
 
 class InboxInvitesReceivedScreen extends StatelessWidget {
   const InboxInvitesReceivedScreen({super.key});
@@ -21,6 +22,7 @@ class InboxInvitesReceivedScreen extends StatelessWidget {
       message: message,
       highlightMessage: true,
       simplifyDate: true,
+      onPressed: () => DebugLogger.warning('TODO: NOT IMPLEMENTED.'),
     );
   }
